custom-elements/form-associated/ElementInternals-NotSupportedError.html

<!DOCTYPE html>
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  class NotFormAssociatedElement extends HTMLElement {}
  customElements.define('my-element1', NotFormAssociatedElement);
  const element = new NotFormAssociatedElement();
  const internals = element.attachInternals();

  assert_throws('NotSupportedError', () => internals.setFormValue(''));
  assert_throws('NotSupportedError', () => internals.form);
  assert_throws('NotSupportedError', () => internals.setValidity({}));
  assert_throws('NotSupportedError', () => internals.willValidate);
  assert_throws('NotSupportedError', () => internals.validity);
  assert_throws('NotSupportedError', () => internals.validationMessage);
  assert_throws('NotSupportedError', () => internals.checkValidity());
  assert_throws('NotSupportedError', () => internals.reportValidity());
  assert_throws('NotSupportedError', () => internals.labels);
}, 'Form-related operations and attributes should throw NotSupportedErrors' +
     ' for non-form-associated custom elements.');
</script>
</body>
